Get agency-level data from the UCR's Offenses Known and Clearances by Arrest data set.

Get agency-level data from the UCR's Offenses Known and Clearances by Arrest data set.

get_agency_crime(ori, key = get_api_key())

Arguments

ori
A string or vector of strings of the 9-character ORI code (unique agency ID) for the desired agency.

key
A string containing your FBI's Crime Data Explorer API key

Value

A data.frame with agency-level UCR crime data for selected agency.

See also

Examples

get_agency_crime("AK0010100", key = "dw3MIpdbpWC8AyD5naDpjXZaEgFenyiGh3y0JP1a")

#>          ori state_abbr year aggravated_assault_cleared
#> 1  AK0010100         AK 1985                        183
#> 2  AK0010100         AK 1986                        241
#> 3  AK0010100         AK 1987                        286
#> 4  AK0010100         AK 1988                        289
#> 5  AK0010100         AK 1989                        337
#> 6  AK0010100         AK 1990                        376
#> 7  AK0010100         AK 1991                        369
#> 8  AK0010100         AK 1992                        565
#> 9  AK0010100         AK 1993                        688
#> 10 AK0010100         AK 1994                        732
#> 11 AK0010100         AK 1995                        728
#> 12 AK0010100         AK 1996                        778
#> 13 AK0010100         AK 1997                        748
#> 14 AK0010100         AK 1998                        659
#> 15 AK0010100         AK 1999                        662
#> 16 AK0010100         AK 2000                        596
#> 17 AK0010100         AK 2001                        704
#> 18 AK0010100         AK 2002                        645
#> 19 AK0010100         AK 2003                        727
#> 20 AK0010100         AK 2004                       1097
#> 21 AK0010100         AK 2005                        896
#> 22 AK0010100         AK 2006                       1213
#> 23 AK0010100         AK 2007                       1155
#> 24 AK0010100         AK 2008                       1223
#> 25 AK0010100         AK 2009                       1134
#> 26 AK0010100         AK 2010                       1129
#> 27 AK0010100         AK 2011                       1003
#> 28 AK0010100         AK 2012                       1107
#> 29 AK0010100         AK 2013                        961
#> 30 AK0010100         AK 2014                       1035
#> 31 AK0010100         AK 2015                       1142
#> 32 AK0010100         AK 2016                       1208
#> 33 AK0010100         AK 2017                       1348
#> 34 AK0010100         AK 2018                       1528
#>    aggravated_assault_actual arson_cleared arson_actual burglary_cleared
#> 1                        606            14          232              177
#> 2                        684            35          189              184
#> 3                        571             7          164              204
#> 4                        669             2           77              139
#> 5                        709             1           39              144
#> 6                        767             0           74              153
#> 7                        838             8          106              168
#> 8                       1217            10          105              164
#> 9                       1410            20          105              192
#> 10                      1529            16          144              181
#> 11                      1462            11           96              176
#> 12                      1297            11           87              155
#> 13                      1191             9           83              190
#> 14                      1056            11           77              150
#> 15                      1106             9           83              223
#> 16                       973            13          101              186
#> 17                      1144            10          103              153
#> 18                      1067            15          141              157
#> 19                      1143            25          128              178
#> 20                      1555            13           82              171
#> 21                      1407             7           76              144
#> 22                      1862            18          106              186
#> 23                      1673            29          145              222
#> 24                      1830            37          106              162
#> 25                      1658            17          111              179
#> 26                      1701            20           86              149
#> 27                      1628            26          126              141
#> 28                      1673            24           97              147
#> 29                      1491            15           74              140
#> 30                      1705            19           81              133
#> 31                      2062            30          106              168
#> 32                      2157            19           69              209
#> 33                      2368            14           62              250
#> 34                      2468            15           46              311
#>    burglary_actual homicide_cleared homicide_actual human_trafficing_cleared
#> 1             2892               10              14                        0
#> 2             2824                8              17                        0
#> 3             2113                7              15                        0
#> 4             1890                9              13                        0
#> 5             1708               12              11                        0
#> 6             2060                7              10                        0
#> 7             2489                9              25                        0
#> 8             2650                9              17                        0
#> 9             1880               13              23                        0
#> 10            2276               14              22                        0
#> 11            2521               19              29                        0
#> 12            2353               16              25                        0
#> 13            1931               21              23                        0
#> 14            1617               11              19                        0
#> 15            1543               10              19                        0
#> 16            1533                9              10                        0
#> 17            1606                8              10                        0
#> 18            1521               19              18                        0
#> 19            1418                9              17                        0
#> 20            1523               14              15                        0
#> 21            1783               10              16                        0
#> 22            1733               12              17                        0
#> 23            1454               17              22                        0
#> 24            1191                7              10                        0
#> 25            1613               12              14                        0
#> 26            1223               10              13                        0
#> 27            1080               13              12                        0
#> 28            1158               12              15                        0
#> 29            1318               12              14                        0
#> 30            1375                9              12                        0
#> 31            1685               21              26                        0
#> 32            2008               20              28                        0
#> 33            2216               28              27                        0
#> 34            2068               14              26                        0
#>    human_trafficing_actual larceny_cleared larceny_actual
#> 1                        0            1908           9154
#> 2                        0            2416          11071
#> 3                        0            1935           9491
#> 4                        0            1744           8340
#> 5                        0            2379           8219
#> 6                        0            2232           8083
#> 7                        0            2824           9967
#> 8                        0            2785          10813
#> 9                        0            2754          10660
#> 10                       0            2752          11717
#> 11                       0            2525          11152
#> 12                       0            2239          10163
#> 13                       0            2239          10083
#> 14                       0            1994           8834
#> 15                       0            1870           8471
#> 16                       0            2162           8799
#> 17                       0            1865           8648
#> 18                       0            2158           9255
#> 19                       0            2062           9091
#> 20                       0            1530           7354
#> 21                       0            1694           8248
#> 22                       0            2040           8543
#> 23                       0            1962           8397
#> 24                       0            1914           7254
#> 25                       0            1889           7835
#> 26                       0            1844           8178
#> 27                       0            1707           7750
#> 28                       0            1868           8554
#> 29                       0            2220           9845
#> 30                       0            1713           9217
#> 31                       0            1603           8962
#> 32                       0            1481          10599
#> 33                       0            1358          10721
#> 34                       0            1305           9498
#>    motor_vehicle_theft_cleared motor_vehicle_theft_actual
#> 1                           72                       1392
#> 2                           88                       1386
#> 3                           94                       1117
#> 4                          151                       1182
#> 5                          172                       1158
#> 6                          204                       1534
#> 7                          160                       1561
#> 8                          170                       1611
#> 9                          179                       1387
#> 10                         342                       2191
#> 11                         317                       2122
#> 12                         210                       1584
#> 13                         176                       1362
#> 14                         226                       1281
#> 15                         203                       1251
#> 16                         163                       1010
#> 17                         171                       1212
#> 18                         220                       1173
#> 19                         214                       1197
#> 20                         192                       1024
#> 21                         231                       1334
#> 22                         243                       1445
#> 23                         234                       1256
#> 24                         147                        766
#> 25                         158                        868
#> 26                         139                        813
#> 27                         126                        625
#> 28                         168                        831
#> 29                         182                        869
#> 30                         143                        939
#> 31                         214                       1154
#> 32                         338                       2042
#> 33                         497                       3104
#> 34                         461                       2823
#>    property_crime_cleared property_crime_actual rape_cleared rape_actual
#> 1                    2157                 13438            0           0
#> 2                    2688                 15281            0           0
#> 3                    2233                 12721            0           0
#> 4                    2034                 11412            0           0
#> 5                    2695                 11085            0           0
#> 6                    2589                 11677            0           0
#> 7                    3152                 14017            0           0
#> 8                    3119                 15074            0           0
#> 9                    3125                 13927            0           0
#> 10                   3275                 16184            0           0
#> 11                   3018                 15795            0           0
#> 12                   2604                 14100            0           0
#> 13                   2605                 13376            0           0
#> 14                   2370                 11732            0           0
#> 15                   2296                 11265            0           0
#> 16                   2511                 11342            0           0
#> 17                   2189                 11466            0           0
#> 18                   2535                 11949            0           0
#> 19                   2454                 11706            0           0
#> 20                   1893                  9901            0           0
#> 21                   2069                 11365            0           0
#> 22                   2469                 11721            0           0
#> 23                   2418                 11107            0           0
#> 24                   2223                  9211            0           0
#> 25                   2226                 10316            0           0
#> 26                   2132                 10214            0           0
#> 27                   1974                  9455            0           0
#> 28                   2183                 10543            0           0
#> 29                   2542                 12032           87         408
#> 30                   1989                 11531           78         392
#> 31                   1985                 11801           69         517
#> 32                   2028                 14649           80         542
#> 33                   2105                 16041           58         607
#> 34                   2077                 14389          123         613
#>    rape_legacy_cleared rape_legacy_actual robbery_cleared robbery_actual
#> 1                   17                193              51            371
#> 2                   26                167              78            330
#> 3                   52                154              58            285
#> 4                   35                159              58            281
#> 5                   15                139              42            272
#> 6                   29                203              48            350
#> 7                   30                264              83            542
#> 8                   22                253              64            484
#> 9                   39                212              90            568
#> 10                  36                198             102            729
#> 11                  39                242             131            777
#> 12                  43                198             115            558
#> 13                  38                174             101            501
#> 14                  30                184              87            373
#> 15                  25                160              97            398
#> 16                  41                195              67            346
#> 17                  48                210              79            384
#> 18                  68                254              79            382
#> 19                  53                244              84            340
#> 20                  71                263              78            331
#> 21                  41                224              84            384
#> 22                  39                248             120            465
#> 23                  59                257             121            453
#> 24                  30                263             195            544
#> 25                  51                282             192            534
#> 26                  41                264             160            454
#> 27                  70                283             154            465
#> 28                  77                303             140            488
#> 29                   0                  0             166            522
#> 30                   0                  0             156            496
#> 31                   0                  0             213            621
#> 32                   0                  0             205            695
#> 33                   0                  0             216            778
#> 34                   0                  0             276            717
#>    violent_crime_cleared violent_crime_actual
#> 1                    261                 1184
#> 2                    353                 1198
#> 3                    403                 1025
#> 4                    391                 1122
#> 5                    406                 1131
#> 6                    460                 1330
#> 7                    491                 1669
#> 8                    660                 1971
#> 9                    830                 2213
#> 10                   884                 2478
#> 11                   917                 2510
#> 12                   952                 2078
#> 13                   908                 1889
#> 14                   787                 1632
#> 15                   794                 1683
#> 16                   713                 1524
#> 17                   839                 1748
#> 18                   811                 1721
#> 19                   873                 1744
#> 20                  1260                 2164
#> 21                  1031                 2031
#> 22                  1384                 2592
#> 23                  1352                 2405
#> 24                  1455                 2647
#> 25                  1389                 2488
#> 26                  1340                 2432
#> 27                  1240                 2388
#> 28                  1336                 2479
#> 29                  1226                 2435
#> 30                  1278                 2605
#> 31                  1445                 3226
#> 32                  1513                 3422
#> 33                  1650                 3780
#> 34                  1941                 3824