Get arrest data from the UCR's Arrests by Age, Sex, and Race data set.

Get arrest data from the UCR's Arrests by Age, Sex, and Race data set.

get_arrest_count(
  ori = NULL,
  state_abb = NULL,
  region = NULL,
  monthly = FALSE,
  key = get_api_key()
)

Arguments

ori
A string or vector of strings of the 9-character ORI code (unique agency ID) for the desired agency.

state_abb
String or vector of strings input for state abbreviation(s) to get data for. If NULL (default) returns national data. If `state_abb` and `region` both have values, will use return `region` input data.

region
String or vector of strings input for region name(s) to get data for. Please run `list_regions()` to see all possible variables.

monthly
If TRUE (not default), returns data as monthly units. Otherwise returns annual data.

key
A string containing your FBI's Crime Data Explorer API key

Value

A data.frame with the number of arrests for each crime-year in the jurisdiction.

Examples

get_arrest_count(ori = "CA0010900")

#>    aggravated_assault all_other_offenses arson burglary curfew
#> 1                1347               5483    79     1442      2
#> 2                1612               5278    77     1403      0
#> 3                1873               6116    77     1445      2
#> 4                1556               6061    62     1625      0
#> 5                1539               6188    68     1526      0
#> 6                1704               6511    50     1160      1
#> 7                1593               5208    63     1138      0
#> 8                1644               4425    53     1229      0
#> 9                1355               8425    48     1039      1
#> 10               1209               8480    24      938      0
#> 11               1279               9209    21      756      0
#> 12               1483               9386    22      721      0
#> 13               1575               7267    14      628      0
#> 14               1490               5315    14      549      0
#> 15               1032               4906    16      440      0
#> 16                946               4945    18      411      2
#> 17               1051               5443    17      474      1
#> 18                902               4388     8      470      1
#> 19                570               3549     4      334      2
#> 20                475               2601     8      227      0
#> 21                609               3548    13      334      0
#> 22                646               4267    14      323      0
#> 23                684               4242    15      434      0
#> 24                553               3735    11      387      0
#> 25                516               2912     5      329      0
#> 26                543               2011    13      286      0
#> 27                478               1471     8      257      0
#> 28                545               2289     5      241      1
#> 29                654               2771     9      219      0
#> 30                664               2951     5      210      3
#> 31                822               2791    14      244      0
#> 32                876               2175    27      209      0
#>    disorderly_conduct  dui drug_grand_total drug_poss_marijuana drug_poss_opium
#> 1                 499 1888             2693                 370             876
#> 2                 491 1521             3874                 198            1504
#> 3                 485 1311             6389                 138            3761
#> 4                 422  983             8028                 126            4640
#> 5                 451  982            12708                 329            6375
#> 6                 462 1086            11380                 437            5162
#> 7                 313 1168            10777                 319            5449
#> 8                 369  987             9695                 327            4842
#> 9                 188  931             7963                 285            3714
#> 10                179  956             7601                 308            3858
#> 11                 84 1250             5417                 303            2332
#> 12                111 1206             5548                 456            2180
#> 13                 98  690             5914                 448            2325
#> 14                104  743             7237                 521            2830
#> 15                 98  519             5914                 426            2707
#> 16                 45  469             5465                 361            2712
#> 17                 60  532             5743                 375            2667
#> 18                 42  413             4655                 282            2017
#> 19                 40  757             3119                 108            1333
#> 20                 22  369             2559                  91            1185
#> 21                 46  313             3517                  74            1474
#> 22                 34  278             4033                  96            1611
#> 23                 70  308             3629                 153            1595
#> 24                153  286             2824                 103            1095
#> 25                160  170             1343                  19             568
#> 26                 78   94             1071                  13             388
#> 27                 10   94             1131                  15             417
#> 28                115  232             1599                  34             535
#> 29                 31  273             1485                  23             437
#> 30                 40  317             1499                  31             427
#> 31                 10  294             1332                  35             406
#> 32                 12  250              824                  39             224
#>    drug_poss_other drug_poss_subtotal drug_poss_synthetic drug_sales_marijuana
#> 1              161               1418                   0                  646
#> 2              169               1877                   0                  181
#> 3              375               4296                   0                  135
#> 4              518               5284                   0                  114
#> 5             1454               8158                   0                  395
#> 6             1107               6706                   0                  556
#> 7              996               6764                   0                  621
#> 8              910               6079                   0                  652
#> 9              817               4816                   0                  650
#> 10             923               5089                   0                  546
#> 11             583               3218                   0                  624
#> 12             650               3286                   0                  768
#> 13             613               3386                   0                  719
#> 14             573               3924                   0                 1061
#> 15             603               3736                   0                  823
#> 16             714               3787                   0                  580
#> 17             764               3806                   0                  736
#> 18             761               3060                   0                  500
#> 19             370               1811                   0                  395
#> 20             379               1655                   0                  283
#> 21             655               2203                   0                  487
#> 22             890               2597                   0                  601
#> 23             741               2489                   0                  486
#> 24             468               1666                   0                  538
#> 25             297                883                   0                  223
#> 26             341                744                   0                  168
#> 27             356                788                   0                  144
#> 28             474               1043                   0                  264
#> 29             485                945                   0                  239
#> 30             452                910                   0                  250
#> 31             514                955                   0                   78
#> 32             334                597                   0                   25
#>    drug_sales_opium drug_sales_other drug_sales_subtotal drug_sales_synthetic
#> 1               612               17                1275                    0
#> 2              1743               26                1950                    0
#> 3              1944               14                2093                    0
#> 4              2621                9                2744                    0
#> 5              4105               50                4550                    0
#> 6              4055               63                4674                    0
#> 7              3322               70                4013                    0
#> 8              2920               44                3616                    0
#> 9              2427               70                3147                    0
#> 10             1929               37                2512                    0
#> 11             1540               35                2199                    0
#> 12             1444               50                2262                    0
#> 13             1752               57                2528                    0
#> 14             2186               66                3313                    0
#> 15             1304               51                2178                    0
#> 16             1060               38                1678                    0
#> 17             1145               56                1937                    0
#> 18             1057               38                1595                    0
#> 19              870               43                1308                    0
#> 20              558               63                 904                    0
#> 21              749               78                1314                    0
#> 22              750               85                1436                    0
#> 23              577               77                1140                    0
#> 24              560               60                1158                    0
#> 25              194               44                 461                    0
#> 26              133               27                 328                    0
#> 27              186               13                 343                    0
#> 28              247               45                 556                    0
#> 29              244               57                 540                    0
#> 30              289               50                 589                    0
#> 31              250               49                 377                    0
#> 32              164               38                 227                    0
#>    drunkness embezzlement forgery fraud gambling_all_others gambling_bookmaking
#> 1      11180           28     341   410                  27                   7
#> 2      11327           26     385   489                  41                  18
#> 3      10096           25     390   472                  19                   6
#> 4       8993           34     318   384                  27                   7
#> 5       9461           40     329   300                  96                   3
#> 6       9060           33     282   271                  49                   4
#> 7       7039           21     325   239                  71                   4
#> 8       5925           16     307   210                  46                   9
#> 9       4764           16     199   137                  86                   1
#> 10      4143            9     197   155                  63                   1
#> 11      3200           18     288   352                  39                   0
#> 12      3289            7     285   370                 112                   1
#> 13      3227            2     272   161                  45                   0
#> 14      3645           20     199   171                  73                   0
#> 15      2978           13     131   151                  28                   0
#> 16      3266           16     215   147                  33                   4
#> 17      3602           20     168   110                  24                   1
#> 18      3296           10     150   151                  22                   0
#> 19      1655            8      95   134                  13                   0
#> 20      1109            5      64    85                  12                   0
#> 21      1423            9      76    89                   6                   0
#> 22      1210           10      78    93                   4                   0
#> 23      1304            5      60    60                  17                   0
#> 24      1095            1      45    34                   0                   0
#> 25       752            2      36    45                   5                   0
#> 26       490            6      47    24                   2                   0
#> 27       307            5      24    25                   3                   0
#> 28       133           10      41    24                   4                   0
#> 29        76           21      36    45                   8                   0
#> 30        23           18      27    24                   6                   0
#> 31         5           12      36    15                   8                   0
#> 32        37           26      25    13                   3                   0
#>    gambling_numbers gambling_total human_trafficking_commercial_sex_traffic
#> 1                 0             34                                        0
#> 2                 0             59                                        0
#> 3                 0             25                                        0
#> 4                 0             34                                        0
#> 5                 0             99                                        0
#> 6                 0             53                                        0
#> 7                 0             75                                        0
#> 8                 0             55                                        0
#> 9                 0             87                                        0
#> 10                0             64                                        0
#> 11                0             39                                        0
#> 12                0            113                                        0
#> 13                0             45                                        0
#> 14                0             73                                        0
#> 15                0             28                                        0
#> 16                0             37                                        0
#> 17                0             25                                        0
#> 18                0             22                                        0
#> 19                0             13                                        0
#> 20                0             12                                        0
#> 21                0              6                                        0
#> 22                0              4                                        0
#> 23                0             17                                        0
#> 24                0              0                                        0
#> 25                0              5                                        0
#> 26                0              2                                        0
#> 27                0              3                                        0
#> 28                0              4                                        0
#> 29                0              8                                        0
#> 30                0              6                                        0
#> 31                0              8                                        0
#> 32                0              3                                        0
#>    human_trafficking_servitude larceny liquor_laws manslaughter
#> 1                            0    3115         102            1
#> 2                            0    3497          46            7
#> 3                            0    3915          42           10
#> 4                            0    4038          44            3
#> 5                            0    3538          48            3
#> 6                            0    3010          31           10
#> 7                            0    2841          52            4
#> 8                            0    2652          42            6
#> 9                            0    1780          20            0
#> 10                           0    1514          34            4
#> 11                           0    1227          41            2
#> 12                           0    1604          61            0
#> 13                           0    1419          54            0
#> 14                           0    1258          54            0
#> 15                           0     843          50            1
#> 16                           0     782         119            3
#> 17                           0     780         147            2
#> 18                           0     750          62            5
#> 19                           0     381          19            3
#> 20                           0     324          65            2
#> 21                           0     357          40            2
#> 22                           0     430          16            1
#> 23                           0     434          35            1
#> 24                           0     319          18            0
#> 25                           0     257           5            0
#> 26                           0     252          10            0
#> 27                           0     234           4            2
#> 28                           0     233          13            0
#> 29                           0     171          18            1
#> 30                           0     131           5            3
#> 31                           0      88           5            2
#> 32                           0     121           9            5
#>    motor_vehicle_theft murder offense_against_family prostitution_total
#> 1                 1014    121                     25               1002
#> 2                 1230    127                     28                689
#> 3                 1576    125                     14                909
#> 4                 2062    131                     28                589
#> 5                 1908    123                     13                545
#> 6                 1685    165                      9                475
#> 7                 1416    151                     23                522
#> 8                 1433    116                     10                395
#> 9                 1284     87                     13                444
#> 10                1174     48                     13                419
#> 11                1037     54                     14               1199
#> 12                1003     52                      4               1404
#> 13                 861     29                      7               1400
#> 14                 900     40                      6               1049
#> 15                 885     39                      5                520
#> 16                 960     37                      8                494
#> 17                 981     44                      7                586
#> 18                 801     53                      7                507
#> 19                 582     30                     19                180
#> 20                 433     31                      2                115
#> 21                 573     37                      3                249
#> 22                 488     19                      7                255
#> 23                 388     35                      6                549
#> 24                 288     31                      3                464
#> 25                 279     19                      4                199
#> 26                 312     22                      4                170
#> 27                 261     16                      2                171
#> 28                 241     26                      4                237
#> 29                 469     32                      3                496
#> 30                 545     30                      1                586
#> 31                 560     45                      2                481
#> 32                 430     23                      0                382
#>    prostitution_assisting prostitution_performing prostitution_purchasing rape
#> 1                       0                       0                       0  132
#> 2                       0                       0                       0  163
#> 3                       0                       0                       0  167
#> 4                       0                       0                       0  138
#> 5                       0                       0                       0  144
#> 6                       0                       0                       0  170
#> 7                       0                       0                       0  170
#> 8                       0                       0                       0  143
#> 9                       0                       0                       0   98
#> 10                      0                       0                       0   92
#> 11                      0                       0                       0   77
#> 12                      0                       0                       0   88
#> 13                      0                       0                       0   81
#> 14                      0                       0                       0   90
#> 15                      0                       0                       0   55
#> 16                      0                       0                       0   59
#> 17                      0                       0                       0   38
#> 18                      0                       0                       0   29
#> 19                      0                       0                       0   24
#> 20                      0                       0                       0   23
#> 21                      0                       0                       0   26
#> 22                      0                       0                       0   38
#> 23                      0                       0                       0   40
#> 24                      0                       0                       0   39
#> 25                      0                       0                       0   24
#> 26                      0                       0                       0   38
#> 27                      0                       0                       0   29
#> 28                      0                       0                       0   31
#> 29                      0                       0                       0   35
#> 30                      0                       0                       0   51
#> 31                      0                       0                       0   51
#> 32                      0                       0                       0   30
#>    robbery other_sex_offenses simple_assault stolen_property suspicion vagrancy
#> 1      700                569           2297             757         0       76
#> 2      692                636           3117             644         0      110
#> 3      706                611           3204             592         0      123
#> 4      676                369           3087             710         0       97
#> 5      777                354           2775             624         0       82
#> 6      846                482           2747             654         0       98
#> 7      810                404           2470             599         0       56
#> 8      835                390           2283             473         0       77
#> 9      691                378           2001             426         0      131
#> 10     719                256           1753             408         0       67
#> 11     638                253           1496             338         0      180
#> 12     617                373           1597             319         0      305
#> 13     472                279           1583             235         0      228
#> 14     478                236           1589             247         0      153
#> 15     319                202           1341             190         0       44
#> 16     382                249           1427             166         0       26
#> 17     444                262           1756             206         0       42
#> 18     386                185           1582             136         0       35
#> 19     287                383           1268             110         0       18
#> 20     273                420           1261              91         0        6
#> 21     470                376           1789             118         0       12
#> 22     534                268           2050             161         0       11
#> 23     537                271           2237             111         0       43
#> 24     461                161           1953              95         0       14
#> 25     342                235           1665              79         0       16
#> 26     321                192           1867              73         0       14
#> 27     334                138           1801              89         0       17
#> 28     294                266           2189             114         0       18
#> 29     265                216           2305             129         0       10
#> 30     316                177           2129             154         0       30
#> 31     336                183           1978             128         0       15
#> 32     415                107           1925             127         0        6
#>    vandalism weapons year
#> 1        715     828 1985
#> 2        810     852 1986
#> 3        803     864 1987
#> 4        788     823 1988
#> 5        749     895 1989
#> 6        762    1048 1990
#> 7        699    1182 1991
#> 8        646    1110 1992
#> 9        559    1129 1993
#> 10       489     968 1994
#> 11       312     592 1996
#> 12       321     698 1997
#> 13       293     475 1998
#> 14       319     391 1999
#> 15       245     298 2000
#> 16       213     282 2001
#> 17       237     299 2002
#> 18       212     285 2003
#> 19       158     254 2005
#> 20       256     288 2006
#> 21       538     458 2007
#> 22       364     570 2008
#> 23       287     556 2009
#> 24       202     491 2010
#> 25       181     366 2011
#> 26       325     231 2012
#> 27       447     204 2013
#> 28       635     241 2014
#> 29       726     346 2015
#> 30       726     412 2016
#> 31       626     435 2017
#> 32       538     376 2018

get_arrest_count()

#> Response [https://api.usa.gov/crime/fbi/sapi/api/data/arrest/agencies/offense/national/all/1985/2018?API_KEY=dw3MIpdbpWC8AyD5naDpjXZaEgFenyiGh3y0JP1a]
#>   Date: 2020-02-09 01:07
#>   Status: 500
#>   Content-Type: application/json;charset=UTF-8
#>   Size: 214 B
#> {
#>   "timestamp" : "2020-02-09T01:07:40.953+0000",
#>   "status" : 500,
#>   "error" : "Internal Server Error",
#>   "message" : "No message available",
#>   "path" : "/api/data/arrest/agencies/offense/national/all/1985/2018"