Get state-, region- or national-level estimated arson data from the UCR's Offenses Known and Clearances by Arrest data set.

Get state-, region- or national-level estimated arson data from the UCR's Offenses Known and Clearances by Arrest data set.

get_estimated_arson(state_abb = NULL, region = NULL,
  key = get_api_key())

Arguments

state_abb
String or vector of strings input for state abbreviation(s) to get data for. If NULL (default) returns national data. If `state_abb` and `region` both have values, will use return `region` input data.

region
String or vector of strings input for region name(s) to get data for. Please run `list_regions()` to see all possible variables.

key
A string containing your FBI's Crime Data Explorer API key

Value

A data.frame with state-level estimated UCR arson data for selected state

See also

Examples

get_estimated_arson("CA")

#>    year state_id state_abbr reported unfounded actual cleared juvenile_cleared
#> 1  1979        6         CA    18421       520  17901    2356             1320
#> 2  1980        6         CA    28918       442  28476    2906             1266
#> 3  1981        6         CA    24865       345  24520    2787             1177
#> 4  1982        6         CA    20640       377  20263    2437              995
#> 5  1983        6         CA    18177       528  17649    2377              925
#> 6  1984        6         CA    20404      1008  19396    2401              969
#> 7  1985        6         CA    20792       380  20412    2426              928
#> 8  1986        6         CA    19835       293  19542    2278              834
#> 9  1987        6         CA    18794       324  18470    2267              826
#> 10 1988        6         CA    19109       305  18804    2146              885
#> 11 1989        6         CA    19372       354  19018    2112              899
#> 12 1990        6         CA    19737       357  19380    2160              880
#> 13 1991        6         CA    19633       545  19088    2222              947
#> 14 1992        6         CA    22430       466  21964    2256              993
#> 15 1993        6         CA    20916       572  20344    2323             1088
#> 16 1994        6         CA    19116       477  18639    2291             1116
#> 17 1995        6         CA    17512       410  17102    2063              986
#> 18 1996        6         CA    18354       402  17952    2048             1022
#> 19 1997        6         CA    16224       341  15883    1909              933
#> 20 1998        6         CA    14314         1  14313       1                0
#> 21 1999        6         CA    14775       325  14450    1696              914
#> 22 2000        6         CA    14621       200  14421    1597              802
#> 23 2001        6         CA    15258        90  15168    1680              854
#> 24 2002        6         CA    14033        79  13954    1540              718
#> 25 2003        6         CA    13645        65  13580    1363              615
#> 26 2004        6         CA    12733        72  12661    1453              617
#> 27 2005        6         CA    12313        60  12253    1478              627
#> 28 2006        6         CA    12771        50  12721    1472              570
#> 29 2007        6         CA    11534        79  11455    1413              572
#> 30 2008        6         CA    10778        80  10698    1289              508
#> 31 2009        6         CA     9220        36   9184    1141              430
#> 32 2010        6         CA     7901        36   7865    1072              383
#> 33 2011        6         CA     7236        36   7200     995              317
#> 34 2012        6         CA     7440        29   7411    1123              291
#> 35 2013        6         CA     7492        34   7458    1201              278
#> 36 2014        6         CA     7185        38   7147    1242              224
#> 37 2015        6         CA     7434        54   7380    1306              196
#> 38 2016        6         CA     7807        41   7766    1323              198
#> 39 2017        6         CA     8688        40   8648    1623              191
#> 40 2018        6         CA     8617        93   8524    1682              126
#>    uninhabited est_damage_value
#> 1          642         86143281
#> 2         1097        155654047
#> 3         1342        177283075
#> 4         1019        159030449
#> 5          874        122941306
#> 6         1059        153169895
#> 7         1056        189913113
#> 8         1167        168087048
#> 9         1161        139006919
#> 10        1146        175765168
#> 11        1285        201129198
#> 12        1403        449740074
#> 13        1368        161084746
#> 14        1575        729352664
#> 15        1347        703495532
#> 16        1398        150462584
#> 17        1798        180269674
#> 18        1279        138470391
#> 19        1211        107244651
#> 20           0             3305
#> 21        1108        137036910
#> 22        1031        150311121
#> 23         994        168414328
#> 24         905        148866866
#> 25         879        179641117
#> 26         777        154126613
#> 27         772        185623918
#> 28         849        173300332
#> 29         783        222785642
#> 30         763        178064481
#> 31         719        158625251
#> 32         689        160555065
#> 33         590         77610608
#> 34         608        100113953
#> 35         583        100271910
#> 36         604        140454793
#> 37         586        137058749
#> 38         583        104359368
#> 39         684        261124686
#> 40         631        172672061