Get state- or national-level estimated crime counts from the UCR's Offenses Known and Clearances by Arrest data set.

Get state- or national-level estimated crime counts from the UCR's Offenses Known and Clearances by Arrest data set.

get_estimated_crime(state_abb = NULL, key = get_api_key())

Arguments

state_abb
String or vector of strings input for state abbreviation(s) to get data for. If NULL (default) returns national data.

key
A string containing your FBI's Crime Data Explorer API key

Value

A data.frame with state-level estimated UCR crime data for selected state

See also

Examples

get_estimated_crime("CA")

#>    state_abbr      state state_id year population violent_crime homicide
#> 1          CA California        6 2018   39557045        176982     1739
#> 2          CA California        6 2017   39399349        178597     1830
#> 3          CA California        6 2016   39296476        174796     1930
#> 4          CA California        6 2015   38993940        166883     1861
#> 5          CA California        6 2014   38792291        153763     1700
#> 6          CA California        6 2013   38431393        154739     1746
#> 7          CA California        6 2012   37999878        160944     1884
#> 8          CA California        6 2011   37683933        154943     1792
#> 9          CA California        6 2010   37338198        164133     1809
#> 10         CA California        6 2009   36961664        174934     1972
#> 11         CA California        6 2008   36756666        185329     2142
#> 12         CA California        6 2007   36553215        191561     2262
#> 13         CA California        6 2006   36457549        194483     2486
#> 14         CA California        6 2005   36154147        190178     2503
#> 15         CA California        6 2004   35842038        189175     2392
#> 16         CA California        6 2003   35462712        205551     2407
#> 17         CA California        6 2002   35001986        208388     2395
#> 18         CA California        6 2001   34600463        212867     2206
#> 19         CA California        6 2000   33871648        210531     2079
#> 20         CA California        6 1999   33145121        207879     2005
#> 21         CA California        6 1998   32667000        229883     2171
#> 22         CA California        6 1997   32268000        257582     2579
#> 23         CA California        6 1996   31878000        274996     2916
#> 24         CA California        6 1995   31589000        305154     3531
#> 25         CA California        6 1994   31431000        318395     3703
#> 26         CA California        6 1993   31211000        336381     4096
#> 27         CA California        6 1992   30867000        345624     3921
#> 28         CA California        6 1991   30380000        331122     3859
#> 29         CA California        6 1990   29760021        311051     3553
#> 30         CA California        6 1989   29063000        284136     3158
#> 31         CA California        6 1988   28168000        261912     2936
#> 32         CA California        6 1987   27663000        253943     2924
#> 33         CA California        6 1986   26981000        248370     3038
#> 34         CA California        6 1985   26365000        201763     2770
#> 35         CA California        6 1984   25622000        195589     2717
#> 36         CA California        6 1983   25174000        194491     2639
#> 37         CA California        6 1982   24724000        201429     2779
#> 38         CA California        6 1981   24159000        208485     3143
#> 39         CA California        6 1980   23532680        210290     3411
#> 40         CA California        6 1979   22696000        184087     2952
#>    rape_legacy rape_revised robbery aggravated_assault property_crime burglary
#> 1           NA        15505   54326             105412         941618   164632
#> 2           NA        14724   56625             105418         987063   176679
#> 3        13702        13702   54789             104375        1002070   188304
#> 4         9341        12811   52862              99349        1024914   197404
#> 5         8389        11578   48681              91804         947193   202669
#> 6         7464        10324   53640              89029        1018907   232058
#> 7         7837           NA   56521              94702        1049465   245767
#> 8         7665           NA   54291              91195         973822   230075
#> 9         8331           NA   58116              95877         981939   228857
#> 10        8713           NA   64093             100156        1009433   230198
#> 11        8903           NA   69388             104896        1080996   237811
#> 12        9046           NA   70706             109547        1112510   237850
#> 13        9235           NA   71142             111620        1159988   247230
#> 14        9392           NA   63622             114661        1200531   250521
#> 15        9615           NA   61768             115400        1227194   245903
#> 16        9994           NA   63770             129380        1215086   242274
#> 17       10198           NA   64968             130827        1176484   238428
#> 18        9960           NA   64614             136087        1134189   232273
#> 19        9785           NA   60249             138418        1056183   222293
#> 20        9363           NA   60039             136472        1053285   223814
#> 21        9782           NA   68782             149148        1188791   269012
#> 22       10189           NA   81468             163346        1312367   299240
#> 23       10244           NA   94222             167614        1385135   312212
#> 24       10554           NA  104611             186458        1536830   353895
#> 25       10984           NA  112160             191548        1622102   384257
#> 26       11766           NA  126436             194083        1678884   414182
#> 27       12761           NA  130897             198045        1716137   427491
#> 28       12896           NA  124939             189428        1726391   424656
#> 29       12688           NA  112208             182602        1654186   400392
#> 30       11966           NA   96431             172581        1681516   410468
#> 31       11780           NA   86141             161055        1607180   407631
#> 32       12109           NA   83341             155569        1545928   419969
#> 33       12119           NA   92512             140701        1576299   457698
#> 34       11421           NA   86387             101185        1516710   448506
#> 35       11702           NA   83924              97246        1461731   443094
#> 36       12093           NA   85826              93933        1486487   460460
#> 37       12529           NA   91988              94133        1599827   499466
#> 38       13566           NA   93781              97995        1625302   540806
#> 39       13693           NA   90420             102766        1633042   545138
#> 40       12239           NA   75767              93129        1511021   496310
#>    larceny motor_vehicle_theft arson
#> 1   621775              155211  8524
#> 2   642019              168365  8648
#> 3   637010              176756  7766
#> 4   656517              170993  7380
#> 5   592673              151851  7147
#> 6   621557              165292  7458
#> 7   635090              168608  7411
#> 8   596905              146842  7200
#> 9   600558              152524  7865
#> 10  615402              163833  9184
#> 11  650656              192529 10698
#> 12  654526              220134 11455
#> 13  669103              243655 12721
#> 14  692467              257543 12253
#> 15  728687              252604 12661
#> 16  731486              241326 13580
#> 17  715692              222364 13954
#> 18  697739              204177 15168
#> 19  651855              182035 14421
#> 20  660991              168480 14450
#> 21  724262              195517 14313
#> 22  784405              228722 15883
#> 23  830457              242466 17952
#> 24  902456              280479 17102
#> 25  929640              308205 18639
#> 26  945407              319295 20344
#> 27  968534              320112 21964
#> 28  986120              315615 19088
#> 29  951580              302214 19380
#> 30  972603              298445 19018
#> 31  933636              265913 18804
#> 32  896335              229624 18470
#> 33  913004              205597 19542
#> 34  890967              177237 20412
#> 35  857328              161309 19396
#> 36  867123              158904 17649
#> 37  935831              164530 20263
#> 38  921939              162557 24520
#> 39  913070              174834 28476
#> 40  847148              167563 17901