Get agency-, state-, region-, or national-level police staffing data.

Get agency-, state-, region-, or national-level police staffing data.

get_police_employment(
  ori = NULL,
  state_abb = NULL,
  region = NULL,
  key = get_api_key()
)

Arguments

ori
A string or vector of strings of the 9-character ORI code (unique agency ID) for the desired agency.

state_abb
String or vector of strings input for state abbreviation(s) to get data for. If NULL (default) returns national data. If `state_abb` and `region` both have values, will use return `region` input data.

region
String or vector of strings input for region name(s) to get data for. Please run `list_regions()` to see all possible variables.

key
A string containing your FBI's Crime Data Explorer API key

Value

A data.frame with columns for annual number of employees and officers (also broken up by gender).

Examples

if (FALSE) {
# Gets only Oakland Police Department in California
get_police_employment("CA0010900")

# Gets California state-level estimates
get_police_employment(state_abb = "CA")

# Gets national-level estimates
get_police_employment()
}