Return a vector of all offenses in the NIBRS data.

Return a vector of all offenses in the NIBRS data.

list_nibrs_offenses()

Value

A vector containing all permissible offenses to input to `get_nibrs_offenses()`, `get_nibrs_victim()`, and `get_nibrs_offender()`.`

See also

Examples

list_nibrs_offenses()

#>  [1] "aggravated-assault"                                
#>  [2] "all-offenses"                                      
#>  [3] "all-other-larceny"                                 
#>  [4] "all-other-offenses"                                
#>  [5] "animal-cruelty"                                    
#>  [6] "arson"                                             
#>  [7] "assault-offenses"                                  
#>  [8] "assisting-or-promoting-prostitution"               
#>  [9] "bad-checks"                                        
#> [10] "betting"                                           
#> [11] "bribery"                                           
#> [12] "burglary-breaking-and-entering"                    
#> [13] "counterfeiting-forgery"                            
#> [14] "credit-card-automated-teller-machine-fraud"        
#> [15] "crime-against-person"                              
#> [16] "crime-against-property"                            
#> [17] "crime-against-society"                             
#> [18] "curfew-loitering-vagrancy-violations"              
#> [19] "destruction-damage-vandalism-of-property"          
#> [20] "driving-under-the-influence"                       
#> [21] "drug-equipment-violations"                         
#> [22] "drug-violations"                                   
#> [23] "drugs-narcotic-offenses"                           
#> [24] "drunkenness"                                       
#> [25] "embezzlement"                                      
#> [26] "extortion-blackmail"                               
#> [27] "false-pretenses-swindle-confidence-game"           
#> [28] "fondling"                                          
#> [29] "fraud-offenses"                                    
#> [30] "gambling-equipment-violation"                      
#> [31] "gambling-offenses"                                 
#> [32] "hacking-computer-invasion"                         
#> [33] "homicide-offenses"                                 
#> [34] "human-trafficking-commerical-involuntary-servitude"
#> [35] "human-trafficking-commerical-sex-acts"             
#> [36] "human-trafficking-offenses"                        
#> [37] "identity-theft"                                    
#> [38] "impersonation"                                     
#> [39] "incest"                                            
#> [40] "intimidation"                                      
#> [41] "justifiable-homicide"                              
#> [42] "kidnapping-abduction"                              
#> [43] "larceny-theft-offenses"                            
#> [44] "liquor-law-violations"                             
#> [45] "motor-vehicle-theft"                               
#> [46] "murder-and-nonnegligent-manslaughter"              
#> [47] "negligent-manslaughter"                            
#> [48] "operating-promoting-assiting-gambling"             
#> [49] "peeping-tom,pocket-picking"                        
#> [50] "pornography-obscence-material"                     
#> [51] "prostitution"                                      
#> [52] "prostitution-offenses"                             
#> [53] "purchasing-prostitution"                           
#> [54] "purse-snatching"                                   
#> [55] "rape"                                              
#> [56] "robbery"                                           
#> [57] "sex-offenses"                                      
#> [58] "sex-offenses-non-forcible"                         
#> [59] "sexual-assult-with-an-object"                      
#> [60] "shoplifting"                                       
#> [61] "simple-assault"                                    
#> [62] "sodomy"                                            
#> [63] "sports-tampering"                                  
#> [64] "statutory-rape"                                    
#> [65] "stolen-property-offenses"                          
#> [66] "theft-from-building"                               
#> [67] "theft-from-coin-operated-machine-or-device"        
#> [68] "theft-from-motor-vehicle"                          
#> [69] "theft-of-motor-vehicle-parts-or-accessories"       
#> [70] "weapon-law-violation"                              
#> [71] "welfare-fraud"                                     
#> [72] "wire-fraud"